+++ src/main/java/site/codemonster/comon/global/images/migration/TeamImageUrlMigrationRunner.java
@@ -1,4 +1,4 @@
-package site.codemonster.comon.global.images.component;
+package site.codemonster.comon.global.images.migration;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -12,7 +12,6 @@ import java.util.List;
 import site.codemonster.comon.domain.team.entity.Team;
 import site.codemonster.comon.domain.team.repository.TeamRepository;
 import site.codemonster.comon.global.images.enums.ImageConstant;
-import site.codemonster.comon.global.util.convertUtils.ImageFieldConvertUtils;
 
 @Slf4j
 @Component
@@ -21,7 +20,7 @@ import site.codemonster.comon.global.util.convertUtils.ImageFieldConvertUtils;
 public class TeamImageUrlMigrationRunner implements CommandLineRunner {
 
     private final TeamRepository teamRepository;
-    private final ImageFieldConvertUtils imageFieldConvertUtils;
+    private final ImageMigrationUtils imageFieldConvertUtils;
 
     @Transactional
     @Override
